Refactor

+++ src/main/java/com/CBConverter/repository/CurrencyRepository.java
@@ -1,7 +1,6 @@
 package com.CBConverter.repository;
 
-import com.CBConverter.domain.Currency;
-import com.CBConverter.domain.History;
+import com.CBConverter.entities.Currency;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 
@@ -9,7 +8,7 @@ public interface CurrencyRepository extends CrudRepository<Currency, Long> {
 
     @Query(value = "select c.value from currency c where c.char_code = ?1",
             nativeQuery = true)
-    float findValueByChar_code(String charCode);
+    double findValueByChar_code(String charCode);
 
     @Query(value = "select c.nominal from currency c where c.char_code = ?1",
             nativeQuery = true)
